requirements.txt
pandas>=2.2.2
transformers>=4.50.0
torch>=2.6.0
attrs>=25.3.0
sentence-transformers>=3.4.1
scikit-learn>=1.6.1
wordcloud>=1.9.4
faiss-cpu>=1.10.0